+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 1.3.6 (unreleased)
 ------------------
 
-- #2308 Validate non-future and non-past dates on add sample (#2307 port)
+- #2308 Rely on fields when validating values on sample creation (#2307 port)
 - #2249 Fix APIError: None is not supported (upgrade 1.3.6)
 - #2198 Use portal as relative path for sticker icons (#2197 port)
 - #2108 Replace ParentAnalysisRequest ReferenceField by UIDReferenceField (#2069 port)